requirements.txt
streamlit
face-recognition
opencv-python
numpy
Pillow
pandas

